drivers/fs: BlockDevice lba is uint64_t; ext4 composes 48-bit phys block (documented cap)

+++ arch/x86/drivers/AtaBlockDevice.h
@@ -9,8 +9,8 @@ class AtaBlockDevice: public BlockDevice {
 	const char* devName;
 public:
 	AtaBlockDevice(const char* name);
-	int readSectors(unsigned lba, unsigned count, void* buf);
-	int writeSectors(unsigned lba, unsigned count, const void* buf);
+	int readSectors(uint64_t lba, unsigned count, void* buf);
+	int writeSectors(uint64_t lba, unsigned count, const void* buf);
 	unsigned sectorSize();
 	const char* name();
 };
